requirements.txt
# Core dependencies
python-telegram-bot==20.7
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
asyncpg==0.29.0
pydantic==2.5.0

# AI and Translation
aiohttp==3.9.1
openai==1.3.7
requests==2.31.0

# Database and caching
redis==5.0.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
